First version of reworked frontend
