+++ frontend/src/i18n.ts
@@ -15,6 +15,8 @@ i18n
                     index_title: "Add passport",
                     index_header: "Add passport",
                     index_explanation: "Add your passport details in your Yivi app.",
+                    index_store_update_title: "Keep your app up to date",
+                    index_store_update_description: "Install the latest version of the Yivi app to get new features and security fixes. Pick your preferred store below.",
                     index_start: "Add passport",
                     index_error: "Something went wrong please try again.",
                     start_app_title: "Download VC MRTD App",
@@ -26,6 +28,8 @@ i18n
                     index_title: "Paspoort toevoegen",
                     index_header: "Paspoort toevoegen",
                     index_explanation: "Zet je paspoortgegevens in je Yivi-app.",
+                    index_store_update_title: "Houd je app up-to-date",
+                    index_store_update_description: "Installeer de nieuwste versie van de Yivi-app voor nieuwe functies en beveiligingsupdates. Kies hieronder je voorkeurswinkel.",
                     index_start: "Paspoort toevoegen",
                     index_error: "Er is iets misgegaan, probeer het opnieuw.",
                     start_app_title: "Download VC MRTD App",
